Move `Cache` & `Item` to its own file
